converted a bunch more to generics -- renamed archive.File to archive.ArchiveFile to not clash with java.io.File, updated HTTPHeaderName to be an enum and have more values, ConstantHTTPHeaderValue to have a few more values, added a SimpleHTTPHeaderValue that can wrap a String, updated IncompleteFileManager's serialization/deserialization to use PutFields and GetFields so the variables never contain values that are wrong.

+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/Statistics.java
@@ -49,7 +49,7 @@ public class Statistics {
      * @see calculateFractionalUptime
      */
     public int calculateDailyUptime() {
-        return (int)(calculateFractionalUptime()*(float)SECONDS_PER_DAY);
+        return (int)(calculateFractionalUptime()*SECONDS_PER_DAY);
     }
 
     /**
